[onert/cl_common] Introduce cl-common library
This commit introduces cl-common backend library.
It will be used by gpu-cl backend and acl-common library.
